requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
coacd==1.0.5
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
dataclasses-json==0.6.7
decorator==5.1.1
Deprecated==1.2.15
dill==0.3.9
etils==1.11.0
filelock==3.16.1
fonttools==4.55.3
freetype-py==2.5.1
fsspec==2024.12.0
genesis-world==0.2.0
glfw==2.8.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_resources==6.4.5
Jinja2==3.1.4
kiwisolver==1.4.7
lazy_loader==0.4
libigl==2.5.1
llvmlite==0.43.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
matplotlib==3.10.0
mdurl==0.1.2
moviepy==2.1.1
mpmath==1.3.0
mujoco==3.2.5
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
opencv-python==4.10.0.84
OpenEXR==3.3.2
packaging==24.2
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
pooch==1.8.2
proglog==0.1.10
psutil==6.1.1
pycollada==0.8
pydantic==2.7.1
pydantic_core==2.18.2
PyGEL3D==0.5.2
pyglet==2.0.20
pygltflib==1.16.0
Pygments==2.18.0
pymeshlab==2023.12.post2
pyobjc-core==10.3.2
pyobjc-framework-Cocoa==10.3.2
PyOpenGL==3.1.7
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyvista==0.44.2
requests==2.32.3
rich==13.9.4
scikit-image==0.25.0
scipy==1.14.1
scooby==0.10.0
screeninfo==0.8.1
setuptools==75.1.0
six==1.17.0
sympy==1.13.3
taichi==1.7.2
tenacity==9.0.0
tetgen==0.6.4
tifffile==2024.12.12
torch==2.4.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
vtk==9.3.1
wheel==0.44.0
wrapt==1.17.0
zipp==3.21.0
